fix(conn): deadlock when async_connect with 0-RTT
msquic does conn callback with StreamsAvailable event in the same code
path of set conn param session ticket.

This causes deadlock when callback handler wait for quicer async_connect to
release the lock of c_ctx when it is called with an opened handle.

solution is to either do set session ticket in open_connection
Or set session ticket in async_connect without an opened handle

callstack:
    at /home/ubuntu/repo/quic/msquic/src/core/stream_set.c:343