Backport/william/kill switch (#379)
* feat: runtime envvar kill switch QUICER_NO_NIF

By setting QUICER_NO_NIF=1,
NIF module (quicer_nif) will be loaded successfully, and quicer
application could be started but NONE of the fuctions can work.

This is used in some deployment that the quicer functions are not
required/desired but quicer app is embedded in release that requires
it to start successfully.

* chore: make hank happy.

* chore: use QUICER_SKIP_NIF_LOAD for kill switch